backend/requirements.txt
Flask==3.0.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
